ecr-discord-bot/main_cloud_function_python/requirements.txt
pynacl==1.5.0
pyjwt==2.7.0
cryptography==3.3.1